Fixed slight item transformer error caused by round off re #194

+++ library/src/main/java/com/yarolegovich/discretescrollview/DSVOrientation.java
@@ -42,7 +42,7 @@ public enum DSVOrientation {
 
         void offsetChildren(int amount, RecyclerViewProxy lm);
 
-        float getDistanceFromCenter(Point center, int viewCenterX, int viewCenterY);
+        float getDistanceFromCenter(Point center, float viewCenterX, float viewCenterY);
 
         boolean isViewVisible(Point center, int halfWidth, int halfHeight, int endBound, int extraSpace);
 
@@ -104,7 +104,7 @@ public enum DSVOrientation {
         }
 
         @Override
-        public float getDistanceFromCenter(Point center, int viewCenterX, int viewCenterY) {
+        public float getDistanceFromCenter(Point center, float viewCenterX, float viewCenterY) {
             return viewCenterX - center.x;
         }
 
@@ -165,7 +165,7 @@ public enum DSVOrientation {
         }
 
         @Override
-        public float getDistanceFromCenter(Point center, int viewCenterX, int viewCenterY) {
+        public float getDistanceFromCenter(Point center, float viewCenterX, float viewCenterY) {
             return viewCenterY - center.y;
         }
